[ci_cd] Check for directories existence prior further actions

+++ ci_cd/build_all.sh
@@ -4,8 +4,8 @@ set -x
 PREVIOUSDIR="$(echo $PWD)"
 
 ROOT_SCRIPTDIR="$(realpath $(dirname "$0"))"
-rm -rf ${ROOT_SCRIPTDIR}/../packages
-rm -rf ${ROOT_SCRIPTDIR}/../artifacts
+test -d "${ROOT_SCRIPTDIR}/../packages"  && rm -rf ${ROOT_SCRIPTDIR}/../packages
+test -d "${ROOT_SCRIPTDIR}/../artifacts" && rm -rf ${ROOT_SCRIPTDIR}/../artifacts
 
 mkdir -p ${ROOT_SCRIPTDIR}/../packages
 sh ${ROOT_SCRIPTDIR}/darwin/build.sh
